Add NoteRepository constructor.
Gets the instance of the database for the Application.
Stores the noteDao and uses it to access the List of all Notes, wrapped in a
LiveData so it automatically updates and is Observable.

NoteDao.noteDao() is abstract, but we're able to call it because we used the
Room.databaseBuilder() in the NoteDatabase getInstance() method so Room
generates the implementation of noteDao() method.