Add skeletons for all database operations
Room automatically runs the operations that will update the LiveData on the
background thread so we don't need to worry about threading.

However, for the other database operations (insert, update, etc.) we must handle
the threading ourselves.

In fact, Room doesn't allow database operations on the main thread because it
would block the main thread and make it unresponsive.  Room would crash the app
if we try.